requirements.txt
blinker==1.3
chardet==2.3.0
cloud-init==17.1
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
Django==2.0.1
django-photologue==3.8.1
django-sortedm2m==1.5.0
ExifRead==2.1.2
hibagent==1.0.1
idna==2.0
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
MarkupSafe==0.23
oauthlib==1.0.3
Pillow==5.0.0
prettytable==0.7.2
pyasn1==0.1.9
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyserial==3.0.1
python-apt==1.1.0b1
python-debian==0.1.27
python-systemd==231
pytz==2017.3
PyYAML==3.11
requests==2.9.1
six==1.10.0
ssh-import-id==5.5
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
virtualenv==15.0.1
